Fixed UI navigation + Some QOL
Home fragment now contains code originally in MainActivity
Camera is now portrait mode
Replaced 'Dashboard' with 'Scan'
