+++ ReaderApp/QRCodeReader/app/src/main/AndroidManifest.xml
@@ -25,6 +25,10 @@
             android:name=".DisplayQRCode"
             android:exported="false" />
         <activity
+            android:name="com.journeyapps.barcodescanner.CaptureActivity"
+            android:screenOrientation="fullSensor"
+            tools:replace="screenOrientation" />
+        <activity
             android:name=".MainActivity"
             android:exported="true"
             android:label="@string/app_name">
