+++ ReaderApp/QRCodeReader/app/src/main/java/com/example/qrcodereader/MainActivity.java
@@ -50,56 +50,6 @@ public class MainActivity extends AppCompatActivity {
         NavController navController = Navigation.findNavController(this, R.id.nav_host_fragment_activity_main);
         NavigationUI.setupActionBarWithNavController(this, navController, appBarConfiguration);
         NavigationUI.setupWithNavController(binding.navView, navController);
-
-
-        // I'm working on this part - Duy
-        Button profile_button = findViewById(R.id.profile_button);
-        profile_button.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                Intent intent = new Intent(MainActivity.this, OrganizerEventActivity.class);
-                startActivity(intent);
-            }
-        });
-
-        Button MyEventButton = findViewById(R.id.my_event_button);
-        MyEventButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                AlertDialog.Builder builder = new AlertDialog.Builder(MainActivity.this);
-
-                builder.setTitle("Choose an action");
-
-                // Button to go to AttendeeEventActivity
-                builder.setPositiveButton("Go to Your Event Page (Attendee)", new DialogInterface.OnClickListener() {
-                    public void onClick(DialogInterface dialog, int id) {
-                        Intent intent = new Intent(MainActivity.this, AttendeeEventActivity.class);
-                        startActivity(intent);
-                    }
-                });
-
-                // Button to go to OrganizerEventActivity
-                builder.setNegativeButton("Go to Your Event Page (Organizer)", new DialogInterface.OnClickListener() {
-                    public void onClick(DialogInterface dialog, int id) {
-                        Intent intent = new Intent(MainActivity.this, OrganizerEventActivity.class);
-                        startActivity(intent);
-                    }
-                });
-
-                // Cancel button
-                builder.setNeutralButton("Cancel", new DialogInterface.OnClickListener() {
-                    public void onClick(DialogInterface dialog, int id) {
-                        // User cancelled the dialog
-                        dialog.dismiss();
-                    }
-                });
-
-                AlertDialog dialog = builder.create();
-                dialog.show();
-            }
-        });
-
-
       
         String deviceID = Settings.Secure.getString(getContentResolver(), Settings.Secure.ANDROID_ID);
         User user = new User(deviceID, "Guohui Lin");
