+++ ReaderApp/QRCodeReader/app/src/main/java/com/example/qrcodereader/ui/dashboard/DashboardFragment.java
@@ -14,19 +14,10 @@ import com.example.qrcodereader.CameraFragment;
 import com.example.qrcodereader.databinding.FragmentCameraBinding;
 
 public class DashboardFragment extends CameraFragment {
-
+    /*
+    Code for camera fragment can be found in superclass
+     */
     private FragmentCameraBinding binding;
-
-//    public View onCreateView(@NonNull LayoutInflater inflater,
-//                             ViewGroup container, Bundle savedInstanceState) {
-//        DashboardViewModel dashboardViewModel =
-//                new ViewModelProvider(this).get(DashboardViewModel.class);
-//
-//        binding = FragmentCameraBinding.inflate(inflater, container, false);
-//        View root = binding.getRoot();
-//
-//        return root;
-//    }
 
     @Override
     public void onDestroyView() {
