+++ ReaderApp/QRCodeReader/app/src/main/res/menu/bottom_nav_menu.xml
@@ -9,7 +9,7 @@
     <item
         android:id="@+id/navigation_camera"
         android:icon="@drawable/ic_dashboard_black_24dp"
-        android:title="@string/title_dashboard" />
+        android:title="@string/title_scan" />
 
     <item
         android:id="@+id/navigation_notifications"
