+++ ReaderApp/QRCodeReader/app/src/main/res/navigation/mobile_navigation.xml
@@ -14,7 +14,7 @@
     <fragment
         android:id="@+id/navigation_camera"
         android:name="com.example.qrcodereader.ui.dashboard.DashboardFragment"
-        android:label="@string/title_dashboard"
+        android:label="@string/title_scan"
         tools:layout="@layout/fragment_camera" />
 
     <fragment
