+++ ReaderApp/QRCodeReader/app/src/main/res/values/strings.xml
@@ -1,7 +1,7 @@
 <resources>
     <string name="app_name">QRCodeReader</string>
     <string name="title_home">Home</string>
-    <string name="title_dashboard">Dashboard</string>
+    <string name="title_scan">Scan</string>
     <string name="title_notifications">Notifications</string>
     <string name="UploadProfilePicText">Upload Profile Picture</string>
     <string name="RemoveProfilePictureText">Remove Profile Picture</string>
